Checking all this shit in

+++ Engine/Rendering/Texture.h
@@ -36,7 +36,7 @@ public:
   int GetWidth() const { return m_Width; }
   int GetHeight() const { return m_Height; }
 
-  Vector2 GetSize() const { return Vector2(m_Width, m_Height); }
+  Vector2 GetSize() const { return Vector2((int)m_Width, (int)m_Height); }
 private:
 
   unsigned int m_TextureName;
